requirements.txt
# PES Multi-Task Classification - Reorganized Project
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# BioMedCLIP and CLIP
open_clip_torch>=2.20.0

# LoRA Fine-tuning
peft>=0.7.0

# Data Processing
pandas>=1.5.0
openpyxl>=3.0.0
Pillow>=9.0.0

# ML Utilities
scikit-learn>=1.0.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Progress Bar
tqdm>=4.60.0
